requirements.txt
kaggle
pandas
numpy
scikit-learn
matplotlib
seaborn
joblib
scipy
tensorflow-cpu
fastapi
uvicorn
